chore: cleaning up extensions
