+++ packages/vscode/cpp/package.json
