+++ packages/vscode/csharp/LICENSE.md
@@ -61,11 +61,6 @@ This extension pack installs and manages third-party VSCode extensions. Each ext
   - Publisher: Jorge Serrano
   - License: MIT
   - Marketplace: [View Extension](https://marketplace.visualstudio.com/items?itemName&#x3D;jorgeserrano.vscode-csharp-snippets)
-- **Version Lens** (`pflannery.vscode-versionlens`)
-  - Description: Shows the latest version for each package using code lens
-  - Publisher: pflannery
-  - License: MIT
-  - Marketplace: [View Extension](https://marketplace.visualstudio.com/items?itemName&#x3D;pflannery.vscode-versionlens)
 - **Roslynator** (`josefpihrt-vscode.roslynator`)
   - Description: A collection of 500+ analyzers, refactorings and fixes for C#, powered by Roslyn
   - Publisher: Josef Pihrt
