+++ packages/vscode/generic-essential/package.json
@@ -38,6 +38,7 @@
     "alefragnani.bookmarks",
     "streetsidesoftware.code-spell-checker",
     "mikestead.dotenv",
+    "usernamehw.errorlens",
     "christian-kohler.path-intellisense",
     "gruntfuggly.todo-tree",
     "pflannery.vscode-versionlens",
@@ -58,6 +59,7 @@
     "alefragnani.bookmarks",
     "streetsidesoftware.code-spell-checker",
     "mikestead.dotenv",
+    "usernamehw.errorlens",
     "christian-kohler.path-intellisense",
     "gruntfuggly.todo-tree",
     "pflannery.vscode-versionlens",
