+++ packages/vscode/godot/package.json
