+++ packages/vscode/typescript/CHANGELOG.md
@@ -4,14 +4,12 @@
 
 ### Added
 - Initial release of typescript extension pack for vscode
-- 4 required extensions
+- 2 required extensions
 - 0 optional extensions
 
 ### Extensions Included
 
 #### Required Extensions
-- [TypeScript Importer](https://marketplace.visualstudio.com/items?itemName&#x3D;ms-vscode.vscode-typescript-next) - Automatically searches for TypeScript definitions
-- [TypeScript Snippets](https://marketplace.visualstudio.com/items?itemName&#x3D;ms-vscode.vscode-typescript-next) - Code snippets for TypeScript development
-- [TypeScript Debugger](https://marketplace.visualstudio.com/items?itemName&#x3D;ms-vscode.js-debug) - Built-in TypeScript/JavaScript debugger for VS Code
+- [TypeScript Language Features](https://marketplace.visualstudio.com/items?itemName&#x3D;ms-vscode.vscode-typescript-next) - Enhanced TypeScript language support with auto-import and intellisense
 - [json2ts](https://marketplace.visualstudio.com/items?itemName&#x3D;GregorBiswanger.json2ts) - Convert JSON object to typescript interfaces
