+++ packages/vscodium/cpp/package.json
