+++ packages/vscodium/csharp/CHANGELOG.md
@@ -13,11 +13,11 @@
 - [C#](https://open-vsx.org/extension/ms-dotnettools/csharp) - Base language support for C#
 - [C# Dev Kit](https://open-vsx.org/extension/ms-dotnettools/csdevkit) - Official C# extension from Microsoft
 - [EditorConfig for VS Code](https://open-vsx.org/extension/editorconfig/editorconfig) - EditorConfig Support for Visual Studio Code
-- [Error Lens](https://open-vsx.org/extension/usernamehw/errorlens) - Improve highlighting of errors, warnings and other language diagnostics
+- []() - 
 
 #### Optional Extensions  
 - [C# Extensions](https://open-vsx.org/extension/kreativ-software/csharpextensions) - C# IDE Extensions for VSCode
 - [.NET Core Test Explorer](https://open-vsx.org/extension/formulahendry/dotnet-test-explorer) - Test Explorer for .NET Core (MSTest, xUnit, NUnit)
 - [C# Snippets](https://open-vsx.org/extension/jorgeserrano/vscode-csharp-snippets) - C# Snippets for Visual Studio Code
-- [Version Lens](https://open-vsx.org/extension/pflannery/vscode-versionlens) - Shows the latest version for each package using code lens
+- []() - 
 - [Roslynator](https://open-vsx.org/extension/josefpihrt-vscode/roslynator) - A collection of 500+ analyzers, refactorings and fixes for C#, powered by Roslyn
