+++ packages/vscodium/csharp/README.md
@@ -13,7 +13,7 @@ These extensions are essential for csharp development:
 - **[C#](https://open-vsx.org/extension/ms-dotnettools/csharp)** - Base language support for C#
 - **[C# Dev Kit](https://open-vsx.org/extension/ms-dotnettools/csdevkit)** - Official C# extension from Microsoft
 - **[EditorConfig for VS Code](https://open-vsx.org/extension/editorconfig/editorconfig)** - EditorConfig Support for Visual Studio Code
-- **[Error Lens](https://open-vsx.org/extension/usernamehw/errorlens)** - Improve highlighting of errors, warnings and other language diagnostics
+- **[]()** - 
 
 ### 💡 Additional Extensions (5)
 
@@ -22,7 +22,7 @@ These extensions provide extra functionality and convenience:
 - **[C# Extensions](https://open-vsx.org/extension/kreativ-software/csharpextensions)** - C# IDE Extensions for VSCode
 - **[.NET Core Test Explorer](https://open-vsx.org/extension/formulahendry/dotnet-test-explorer)** - Test Explorer for .NET Core (MSTest, xUnit, NUnit)
 - **[C# Snippets](https://open-vsx.org/extension/jorgeserrano/vscode-csharp-snippets)** - C# Snippets for Visual Studio Code
-- **[Version Lens](https://open-vsx.org/extension/pflannery/vscode-versionlens)** - Shows the latest version for each package using code lens
+- **[]()** - 
 - **[Roslynator](https://open-vsx.org/extension/josefpihrt-vscode/roslynator)** - A collection of 500+ analyzers, refactorings and fixes for C#, powered by Roslyn
 
 ## 🚀 Installation
@@ -168,11 +168,11 @@ After installation, you may want to configure some settings for optimal csharp d
 | [C#](https://open-vsx.org/extension/ms-dotnettools/csharp) | Microsoft | Base language support for C# |
 | [C# Dev Kit](https://open-vsx.org/extension/ms-dotnettools/csdevkit) | Microsoft | Official C# extension from Microsoft |
 | [EditorConfig for VS Code](https://open-vsx.org/extension/editorconfig/editorconfig) | EditorConfig | EditorConfig Support for Visual Studio Code |
-| [Error Lens](https://open-vsx.org/extension/usernamehw/errorlens) | Alexander | Improve highlighting of errors, warnings and other language diagnostics |
+| []() |  |  |
 | [C# Extensions](https://open-vsx.org/extension/kreativ-software/csharpextensions) | Kreativ Software | C# IDE Extensions for VSCode |
 | [.NET Core Test Explorer](https://open-vsx.org/extension/formulahendry/dotnet-test-explorer) | Jun Han | Test Explorer for .NET Core (MSTest, xUnit, NUnit) |
 | [C# Snippets](https://open-vsx.org/extension/jorgeserrano/vscode-csharp-snippets) | Jorge Serrano | C# Snippets for Visual Studio Code |
-| [Version Lens](https://open-vsx.org/extension/pflannery/vscode-versionlens) | pflannery | Shows the latest version for each package using code lens |
+| []() |  |  |
 | [Roslynator](https://open-vsx.org/extension/josefpihrt-vscode/roslynator) | Josef Pihrt | A collection of 500+ analyzers, refactorings and fixes for C#, powered by Roslyn |
 
 ## 🏷️ Categories
@@ -192,11 +192,11 @@ This extension pack is licensed under the **MIT License** - see [LICENSE.md](htt
 | [C#](https://open-vsx.org/extension/ms-dotnettools/csharp) | Microsoft | MIT | Base language support for C# |
 | [C# Dev Kit](https://open-vsx.org/extension/ms-dotnettools/csdevkit) | Microsoft | MIT | Official C# extension from Microsoft |
 | [EditorConfig for VS Code](https://open-vsx.org/extension/editorconfig/editorconfig) | EditorConfig | MIT | EditorConfig Support for Visual Studio Code |
-| [Error Lens](https://open-vsx.org/extension/usernamehw/errorlens) | Alexander | MIT | Improve highlighting of errors, warnings and other language diagnostics |
+| [](https://marketplace.visualstudio.com/items?itemName=) | Unknown | See extension page |  |
 | [C# Extensions](https://open-vsx.org/extension/kreativ-software/csharpextensions) | Kreativ Software | MIT | C# IDE Extensions for VSCode |
 | [.NET Core Test Explorer](https://open-vsx.org/extension/formulahendry/dotnet-test-explorer) | Jun Han | MIT | Test Explorer for .NET Core (MSTest, xUnit, NUnit) |
 | [C# Snippets](https://open-vsx.org/extension/jorgeserrano/vscode-csharp-snippets) | Jorge Serrano | MIT | C# Snippets for Visual Studio Code |
-| [Version Lens](https://open-vsx.org/extension/pflannery/vscode-versionlens) | pflannery | MIT | Shows the latest version for each package using code lens |
+| [](https://marketplace.visualstudio.com/items?itemName=) | Unknown | See extension page |  |
 | [Roslynator](https://open-vsx.org/extension/josefpihrt-vscode/roslynator) | Josef Pihrt | MIT | A collection of 500+ analyzers, refactorings and fixes for C#, powered by Roslyn |
 
 ### Disclaimer
