+++ packages/vscodium/godot/README.md
@@ -20,10 +20,10 @@ These extensions provide extra functionality and convenience:
 - **[Godot Files](https://open-vsx.org/extension/alfish/godot-files)** - Syntax highlighting for Godot files (.tres, .tscn, .godot, etc.)
 - **[Godot Shaders](https://open-vsx.org/extension/arkii/godot-shaders)** - Syntax highlighting and support for Godot shader files
 - **[Godot Snippets](https://open-vsx.org/extension/razoric/gdscript-snippets)** - Code snippets for GDScript and Godot development
-- **[Error Lens](https://open-vsx.org/extension/usernamehw/errorlens)** - Improve highlighting of errors, warnings and other language diagnostics
+- **[]()** - 
 - **[TODO Highlight](https://open-vsx.org/extension/wayou/vscode-todo-highlight)** - Highlight TODOs, FIXMEs, and other annotations in your code
 - **[Rainbow Brackets](https://open-vsx.org/extension/2gua/rainbow-brackets)** - Provide rainbow colors for the round brackets, the square brackets and the squiggly brackets
-- **[Better Comments](https://open-vsx.org/extension/aaron-bond/better-comments)** - Improve your code commenting by annotating with alert, informational, TODOs, and more
+- **[]()** - 
 
 ## 🚀 Installation
 
@@ -236,10 +236,10 @@ After installation, you may want to configure some settings for optimal godot de
 | [Godot Files](https://open-vsx.org/extension/alfish/godot-files) | alfish | Syntax highlighting for Godot files (.tres, .tscn, .godot, etc.) |
 | [Godot Shaders](https://open-vsx.org/extension/arkii/godot-shaders) | arkii | Syntax highlighting and support for Godot shader files |
 | [Godot Snippets](https://open-vsx.org/extension/razoric/gdscript-snippets) | razoric | Code snippets for GDScript and Godot development |
-| [Error Lens](https://open-vsx.org/extension/usernamehw/errorlens) | usernamehw | Improve highlighting of errors, warnings and other language diagnostics |
+| []() |  |  |
 | [TODO Highlight](https://open-vsx.org/extension/wayou/vscode-todo-highlight) | wayou | Highlight TODOs, FIXMEs, and other annotations in your code |
 | [Rainbow Brackets](https://open-vsx.org/extension/2gua/rainbow-brackets) | 2gua | Provide rainbow colors for the round brackets, the square brackets and the squiggly brackets |
-| [Better Comments](https://open-vsx.org/extension/aaron-bond/better-comments) | aaron-bond | Improve your code commenting by annotating with alert, informational, TODOs, and more |
+| []() |  |  |
 
 ## 🏷️ Categories
 
@@ -260,10 +260,10 @@ This extension pack is licensed under the **MIT License** - see [LICENSE.md](htt
 | [Godot Files](https://open-vsx.org/extension/alfish/godot-files) | alfish | MIT | Syntax highlighting for Godot files (.tres, .tscn, .godot, etc.) |
 | [Godot Shaders](https://open-vsx.org/extension/arkii/godot-shaders) | arkii | MIT | Syntax highlighting and support for Godot shader files |
 | [Godot Snippets](https://open-vsx.org/extension/razoric/gdscript-snippets) | razoric | MIT | Code snippets for GDScript and Godot development |
-| [Error Lens](https://open-vsx.org/extension/usernamehw/errorlens) | usernamehw | MIT | Improve highlighting of errors, warnings and other language diagnostics |
+| [](https://marketplace.visualstudio.com/items?itemName=) | Unknown | See extension page |  |
 | [TODO Highlight](https://open-vsx.org/extension/wayou/vscode-todo-highlight) | wayou | MIT | Highlight TODOs, FIXMEs, and other annotations in your code |
 | [Rainbow Brackets](https://open-vsx.org/extension/2gua/rainbow-brackets) | 2gua | MIT | Provide rainbow colors for the round brackets, the square brackets and the squiggly brackets |
-| [Better Comments](https://open-vsx.org/extension/aaron-bond/better-comments) | aaron-bond | MIT | Improve your code commenting by annotating with alert, informational, TODOs, and more |
+| [](https://marketplace.visualstudio.com/items?itemName=) | Unknown | See extension page |  |
 
 ### Disclaimer
 - We **do not guarantee** the functionality, security, or compatibility of included extensions
